really

+++ code/hud/hudconfig.cpp
@@ -1301,7 +1301,7 @@ void hud_config_button_do(int n)
 			break;
 		}
 
-		if (HC_current_file >= HC_preset_filenames.size()) {
+		if (HC_current_file >= (int)HC_preset_filenames.size()) {
 			HC_current_file = 0;
 		} else {
 			HC_current_file++;
